update to gmp-2.0.0

+++ src/TokenSenderCaller.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.8.0;
 
 import {ISenderCaller, ICallee, Utils} from "./IOATS.sol";
-import {IGateway} from "gmp-1.0.0/src/IGateway.sol";
-import {IGmpReceiver} from "gmp-1.0.0/src/IGmpReceiver.sol";
+import {IGateway} from "gmp-2.0.0/src/IGateway.sol";
+import {IGmpReceiver} from "gmp-2.0.0/src/IGmpReceiver.sol";
 
 import {Ownable} from "@openzeppelin/access/Ownable.sol";
 import {ERC20} from "@openzeppelin/token/ERC20/ERC20.sol";
@@ -45,7 +45,7 @@ contract Token is ISenderCaller, IGmpReceiver, Ownable, ERC20Burnable, ERC20Capp
         Default.caldata = caldata;
         bytes memory message = abi.encode(Default);
 
-        return _gateway.estimateMessageCost(networkId, uint16(message.length), gasLimit);
+        return _gateway.estimateMessageCost(networkId, message.length, gasLimit);
     }
 
     /// @inheritdoc ISenderCaller
